requirements.txt
# Core ML/DL
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
transformers>=4.30.0

# Diffusion Models
diffusers>=0.21.0
accelerate>=0.20.0

# Adversarial Attacks
torchattacks>=3.4.0

# Metrics
lpips>=0.1.4
pytorch-fid>=0.3.0
scikit-image>=0.21.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=9.5.0
opencv-python>=4.8.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# ML Utilities
scikit-learn>=1.3.0

# Configuration
PyYAML>=6.0
omegaconf>=2.3.0

# Jupyter/Colab
ipywidgets>=8.0.0
notebook>=7.0.0
